chore: remove parse5-utils
