+++ packages/dev-server-core/package.json
@@ -56,7 +56,6 @@
     "@parse5/tools": "^0.0.2-6",
     "@types/koa": "^2.11.6",
     "@types/ws": "^7.4.0",
-    "@web/parse5-utils": "^1.3.1",
     "chokidar": "^3.4.3",
     "clone": "^2.1.2",
     "es-module-lexer": "^1.0.0",
