+++ packages/dev-server-hmr/tsconfig.json
@@ -11,9 +11,6 @@
   },
   "references": [
     {
-      "path": "../parse5-utils/tsconfig.json"
-    },
-    {
       "path": "../dev-server-core/tsconfig.json"
     }
   ],
