+++ packages/dev-server-import-maps/package.json
@@ -53,7 +53,6 @@
     "@import-maps/resolve": "^1.0.1",
     "@web/dev-server-core": "^0.4.0",
     "@parse5/tools": "^0.0.2-6",
-    "@web/parse5-utils": "^1.3.0",
     "parse5": "^7.1.0",
     "picomatch": "^2.2.2"
   },
