--- packages/parse5-utils/CHANGELOG.md
@@ -1,55 +1,0 @@
-# @web/parse5-utils
-
-## 1.3.1
-
-### Patch Changes
-
-- 18a16bb0: Update `html-minifier-terser`
-
-## 1.3.0
-
-### Minor Changes
-
-- ca749b0e: Update dependency @types/parse5 to v6
-
-## 1.2.2
-
-### Patch Changes
-
-- a07f4aef: Add missing export for prepend utility.
-
-## 1.2.1
-
-### Patch Changes
-
-- abe37741: Allow break lines in comments when checking isHtmlFragment
-
-## 1.2.0
-
-### Minor Changes
-
-- b5af71e3: Ignore comments when checking isHtmlFragment
-
-## 1.1.2
-
-### Patch Changes
-
-- a7c9af6: fix entrypoint for node v10
-
-## 1.1.1
-
-### Patch Changes
-
-- 68b8211: export using direct assignment
-
-## 1.1.0
-
-### Minor Changes
-
-- 3121966: add textcontent helpers
-
-## 1.0.0
-
-### Major Changes
-
-- cd5244e: First setup
