--- packages/parse5-utils/index.d.ts
@@ -1,1 +1,0 @@
-export * from './dist/index';
