--- packages/parse5-utils/index.mjs
@@ -1,103 +1,0 @@
-import cjsEntrypoint from './src/index.js';
-
-const {
-  createDocument,
-  createDocumentFragment,
-  createElement,
-  createScript,
-  createCommentNode,
-  appendChild,
-  insertBefore,
-  setTemplateContent,
-  getTemplateContent,
-  setDocumentType,
-  setDocumentMode,
-  getDocumentMode,
-  detachNode,
-  insertText,
-  insertTextBefore,
-  adoptAttributes,
-  getFirstChild,
-  getChildNodes,
-  getParentNode,
-  getAttrList,
-  getTagName,
-  getNamespaceURI,
-  getTextNodeContent,
-  getCommentNodeContent,
-  getDocumentTypeNodeName,
-  getDocumentTypeNodePublicId,
-  getDocumentTypeNodeSystemId,
-  isTextNode,
-  isCommentNode,
-  isDocumentTypeNode,
-  isElementNode,
-  setNodeSourceCodeLocation,
-  getNodeSourceCodeLocation,
-  updateNodeSourceCodeLocation,
-  isHtmlFragment,
-  hasAttribute,
-  getAttribute,
-  getAttributes,
-  setAttribute,
-  setAttributes,
-  removeAttribute,
-  remove,
-  findNode,
-  findNodes,
-  findElement,
-  findElements,
-  prependToDocument,
-  appendToDocument,
-} = cjsEntrypoint;
-
-export {
-  createDocument,
-  createDocumentFragment,
-  createElement,
-  createScript,
-  createCommentNode,
-  appendChild,
-  insertBefore,
-  setTemplateContent,
-  getTemplateContent,
-  setDocumentType,
-  setDocumentMode,
-  getDocumentMode,
-  detachNode,
-  insertText,
-  insertTextBefore,
-  adoptAttributes,
-  getFirstChild,
-  getChildNodes,
-  getParentNode,
-  getAttrList,
-  getTagName,
-  getNamespaceURI,
-  getTextNodeContent,
-  getCommentNodeContent,
-  getDocumentTypeNodeName,
-  getDocumentTypeNodePublicId,
-  getDocumentTypeNodeSystemId,
-  isTextNode,
-  isCommentNode,
-  isDocumentTypeNode,
-  isElementNode,
-  setNodeSourceCodeLocation,
-  getNodeSourceCodeLocation,
-  updateNodeSourceCodeLocation,
-  isHtmlFragment,
-  hasAttribute,
-  getAttribute,
-  getAttributes,
-  setAttribute,
-  setAttributes,
-  removeAttribute,
-  remove,
-  findNode,
-  findNodes,
-  findElement,
-  findElements,
-  prependToDocument,
-  appendToDocument,
-};
