--- packages/parse5-utils/tsconfig.json
@@ -1,24 +1,0 @@
-// Don't edit this file directly. It is generated by /scripts/update-package-configs.ts
-
-{
-  "extends": "../../tsconfig.node-base.json",
-  "compilerOptions": {
-    "module": "commonjs",
-    "outDir": "./dist",
-    "rootDir": "./src",
-    "composite": true,
-    "allowJs": true,
-    "checkJs": true,
-    "emitDeclarationOnly": true
-  },
-  "references": [],
-  "include": [
-    "src",
-    "types"
-  ],
-  "exclude": [
-    "src/browser",
-    "tests",
-    "dist"
-  ]
-}
