+++ packages/polyfills-loader/package.json
@@ -47,7 +47,6 @@
   "dependencies": {
     "@babel/core": "^7.12.10",
     "@parse5/tools": "^0.0.2-6",
-    "@web/parse5-utils": "^1.3.0",
     "@webcomponents/shadycss": "^1.11.0",
     "@webcomponents/webcomponentsjs": "^2.5.0",
     "abortcontroller-polyfill": "^1.5.0",
