+++ packages/rollup-plugin-html/package.json
@@ -43,7 +43,7 @@
     "html"
   ],
   "dependencies": {
-    "@web/parse5-utils": "^1.3.1",
+    "@parse5/tools": "^0.0.2-6",
     "glob": "^7.1.6",
     "html-minifier-terser": "^7.1.0",
     "@parse5/tools": "^0.0.2-6",
