+++ packages/rollup-plugin-polyfills-loader/tsconfig.json
@@ -11,9 +11,6 @@
   },
   "references": [
     {
-      "path": "../parse5-utils/tsconfig.json"
-    },
-    {
       "path": "../polyfills-loader/tsconfig.json"
     },
     {
