+++ packages/test-runner-webdriver/tsconfig.json
@@ -11,9 +11,6 @@
   },
   "references": [
     {
-      "path": "../parse5-utils/tsconfig.json"
-    },
-    {
       "path": "../browser-logs/tsconfig.json"
     },
     {
