+++ tsconfig.json
@@ -5,9 +5,6 @@
   "files": [],
   "references": [
     {
-      "path": "./packages/parse5-utils/tsconfig.json"
-    },
-    {
       "path": "./packages/polyfills-loader/tsconfig.json"
     },
     {
